backend/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2
django-filter==2.2.0
djangorestframework==3.10.2
idna==2.8
Markdown==3.1.1
node==0.9.24
odict==1.7.0
optional-django==0.1.0
plumber==1.5
psycopg2-binary==2.8.3
pytz==2019.2
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
zope.component==4.5
zope.deferredimport==4.3.1
zope.deprecation==4.4.0
zope.event==4.4
zope.hookable==4.2.0
zope.interface==4.6.0
zope.lifecycleevent==4.3
zope.proxy==4.3.2
